+++ yass2-core/src/Interceptor.kt
@@ -1,9 +1,9 @@
 package ch.softappeal.yass2.core
 
-// function is String (= function name) instead of KFunction because annotation reflection is not multiplatform.
+import kotlin.reflect.KFunction
 
 public typealias Invocation = suspend () -> Any?
-public typealias Interceptor = suspend (function: String, parameters: List<Any?>, invocation: Invocation) -> Any?
+public typealias Interceptor = suspend (function: KFunction<*>, parameters: List<Any?>, invocation: Invocation) -> Any?
 
 public inline operator fun Interceptor.plus(crossinline interceptor: Interceptor): Interceptor =
     { function, parameters, invocation ->